Renaming StatusNotifier to StatusNotifierItem
Yes, this breaks things. It's unfortunate, but we didn't really follow
the GObject convention for naming things, and as a result it made things
a bit awkward/complicated when it comes to introspection, and bindings
to use statusnotifier from other languages (e.g. Perl).

So, let's do this renaming (and of a few macros as well) to make things
better/simpler, and it shouldn't be too much breakage/too hard to fix
in applications, just a bunch of search/replace as every single function
was renamed.

Note that a statusnotifier-compat.h is provided that defines a bunch of
macros and should allow old code to keep working without any changes
required, only this extra include. As a transitional aid of sorts.