Add register-name-on-bus property
According to the spec we should always register a name on the form
"org.kde.StatusNotifierItem-PID-ID" on the bus to register our
StatusNotifierItem and then send said name to the watcher.

However, this causes issues when running in sandbox where bus name
ownership permissions is likely to not be granted for org.kde.*

So one way to deal with this is to instead not register any name but
simply use the unique name of the existing dbus connection (so in the
form :x.xxxx)

Whilst non confirming, what is sent to the watcher remains a valid dbus
name where the StatusNotifierItem object can indeed be found, so it all
works fine in the end.
(In fact, some (KDE) applications already do this for the same reasons.)

This property will allow applications to either ask not to register a
bus name as described, or try to determine whether running in a sandbox
or not and do the right thing accordingly.

Closes #18